requirements.txt
discord.py==2.1.1
dotenv==0.9.9
pillow==11.2.1
requests==2.31.0
